TextSummarization/requirements.txt
gradio
faiss-cpu
sentence-transformers
transformers
